Gas opt

+++ contracts/interfaces/IImpossibleCallee.sol
@@ -2,5 +2,10 @@
 pragma solidity =0.7.6;
 
 interface IImpossibleCallee {
-    function ImpossibleCall(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function ImpossibleCall(
+        address sender,
+        uint256 amount0,
+        uint256 amount1,
+        bytes calldata data
+    ) external;
 }
